tools/cla/requirements.txt
# For stm32_mtcp_proxy
pyserial
